# [SPARK-16907] [SQL] Fix performance regression for parquet table  when vectorized parquet record reader is not being used