@gatorsmile 2016-08-01
Maybe, we can correct the PR description and make it more accurate.

This PR avoids the extra memory copy when the vectorized parquet record reader is not being used for reading a non-partitioned Parquet table. One of the typical case is the parquet table with non atomic types, including null, UDTs, arrays, structs, and maps. Another case is when users set `spark.sql.parquet.enableVectorizedReader` to `false`. 

Is my understanding correct?