@rxin 2016-08-02
Thanks - merging in master/2.0.