@rxin 2016-08-02
Actually, before I merge, can you create a new JIRA ticket and link to that one instead? I don't want to close the old one because I don't know if this issue by itself has completely solved it.