@maver1ck 2016-08-02
@rxin 
I tested this patch.
The result are almost equal to Spark without this patch. (difference is less than 5%)
So maybe it's needed but it doesn't solve my problem.