@maver1ck 2016-08-03
@rxin 
I added some comments to Jira.
I think both problems has solutions right now.