@cloud-fan 2016-08-05
thanks. merging to master and 2.0!